requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.12
psutil==5.9.6
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5